Fix build for Windows

+++ src/sasl_frame_codec.c
@@ -286,7 +286,7 @@ int sasl_frame_codec_encode_frame(SASL_FRAME_CODEC_HANDLE sasl_frame_codec, AMQP
         else if (encoded_size < MIN_FRAME_SIZE || encoded_size > (MIX_MAX_FRAME_SIZE - 8))
         {
             /* Codes_SRS_SASL_FRAME_CODEC_01_034: [If any error occurs during encoding, sasl_frame_codec_encode_frame shall fail and return a non-zero value.] */
-            LogError("SASL frame encoded size out of bounds (%ld)", encoded_size);
+            LogError("SASL frame encoded size out of bounds (%u)", (unsigned int)encoded_size);
             result = MU_FAILURE;
         }
         else
